Update README.md
"Today, I downloaded a data table. That data table is about 'Most Sixes in Test Cricket'. In it, there are many columns like Matches, Innings, Not Out, Runs, Highest Score, Average, Balls Faced, Strike Rate, 100s, 50s, Ducks (or 0s), 4s, and 6s.I want the data for 'Most Sixes'. So, I only need the '6s' column. Looking at this data table, it has both inputs and outputs. So, this is an example of Supervised Learning. In this data table, '6s' is dependent data because it depends on the inputs like Matches, Innings, Runs, Highest Score, etc. So, the '6s' column is dependent data. And when we look at the values inside the '6s' column, they are all numerical data — because they start with numbers like 1, 2, 3, etc. So, this data comes under Numerical Data. These are all my points of view."